Breadcrumb cleanup, confirm delete, consolidate views
Added app.js to public folder and then referenced it in
layout.blade.php to make sure deletes are confirmed.  Consolidated
views - make modules appear on nodes.show and made ports appear on
nodes.modules.show.

+++ resources/views/nodes/index.blade.php
@@ -22,24 +22,23 @@
 			<div class="panel-body">
                                 <table class="table table-hover">
                                     <tr>
-                                        <th>Name</th>
+                                        <th>Node Name</th>
                                         <th>Location</th>
-                                        <th>Management IP</th>
                                         <th>Model</th>
+                                        <th nowrap>Management IP</th>
                                         <th>Notes</th>
-                                        <th>Edit</th>
                                         <th>Delete</th>
                                     </tr>
                                     @foreach ($nodes as $node)
                                         <tr>
-                                            <td> {!! link_to_route('nodes.show', $node->name, $node->id ) !!}</td>
+                                            <td>{!! link_to_route('nodes.show', $node->name, $node->id ) !!}</td>
                                             <td nowrap>{{ $node->location }}</td>
-                                            <td>{{ $node->mgmt_ip }}</td>
                                             <td>{{ $node->model_number }}</td>
+                                            <td>{{ $node->mgmt_ip }}</td>
                                             <td>{{ $node->notes }}</td>
-                                            <td>{!! link_to_route('nodes.edit', 'Edit', $node->id, ['class' => 'btn btn-primary btn-xs']) !!}</td>
                                             <td>
-                                                {!! Form::model($node, ['route' => ['nodes.destroy', $node->id], 'method' => 'delete' ]) !!}
+                                                <!-- {!! Form::model($node, ['route' => ['nodes.destroy', $node->id], 'method' => 'delete' ]) !!} -->
+                                                {!! Form::model($node, ['route' => ['nodes.destroy', $node->id], 'method' => 'delete', 'onsubmit' => 'ConfirmDelete()']) !!}
                                                 {!! Form::button('delete', ['type' => 'submit', 'class' => 'btn btn-primary btn-xs'] ) !!}
                                                 {!! Form::close() !!}    
                                             </td>
